+++ tests/tests.py
@@ -371,6 +371,18 @@ class PhonenumerFieldAppTest(TestCase):
         self.assertIsInstance(models.TestModel.phone, modelfields.PhoneNumberDescriptor)
 
 
+class PhoneNumberFormFieldTest(TestCase):
+    def test_error_message(self):
+        class PhoneNumberForm(forms.Form):
+            number = formfields.PhoneNumberField()
+
+        form = PhoneNumberForm({"number": "invalid"})
+        self.assertIs(form.is_valid(), False)
+        self.assertEqual(
+            form.errors, {"number": ["Enter a valid phone number (e.g. +12125552368)."]}
+        )
+
+
 class RegionPhoneNumberFormFieldTest(TestCase):
     def test_regional_phone(self):
         class PhoneNumberForm(forms.Form):
@@ -391,6 +403,22 @@ class RegionPhoneNumberFormFieldTest(TestCase):
 
         self.assertTrue(
             force_text(cm.exception).startswith("“invalid” is not a valid region code.")
+        )
+
+    def test_error_message_nationalize_example(self):
+        class PhoneNumberForm(forms.Form):
+            number = formfields.PhoneNumberField(region="CA")
+
+        form = PhoneNumberForm({"number": "invalid"})
+        self.assertIs(form.is_valid(), False)
+        self.assertEqual(
+            form.errors,
+            {
+                "number": [
+                    "Enter a valid phone number (e.g. (506) 234-5678) "
+                    "or a number with an international call prefix."
+                ]
+            },
         )